Extract site upgrade steps to separate script
It's important to understand that changes in upgrade.sh are not executed
when they are deployed, since the script is always running as of its
last deployed version. That's why it's important to keep upgrade.sh as
lean as possible. Anything that might change in the future, for example
the details of deploying site files, is better in a separate script.
Note that the separate script should not be renamed, as that would be an
incompatible change, and require multiple deployments to take effect.